# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kitna - Money clarity dashboard for creators and influencer managers
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: List all payments
      parameters:
        - name: deal_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Record a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/{id}:
    patch:
      operationId: updatePayment
      tags: [payments]
      summary: Update a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePaymentBody"
      responses:
        "200":
          description: Payment updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deletePayment
      tags: [payments]
      summary: Delete a payment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Payment deleted
        "404":
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses:
    get:
      operationId: getExpenses
      tags: [expenses]
      summary: List all expenses
      parameters:
        - name: deal_id
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"

    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExpenseBody"
      responses:
        "200":
          description: Expense updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Expense deleted
        "404":
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dealId:
          type: ["integer", "null"]
        amountReceived:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        receivedDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
        dealBrandName:
          type: ["string", "null"]
        dealCampaignName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - amountReceived
        - currency
        - paymentMethod
        - receivedDate
        - tdsApplicable
        - createdAt
        - updatedAt
    CreatePaymentBody:
      type: object
      properties:
        dealId:
          type: ["integer", "null"]
        amountReceived:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        receivedDate:
          type: string
        notes:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
      required:
        - amountReceived
        - paymentMethod
        - receivedDate
    UpdatePaymentBody:
      type: object
      properties:
        dealId:
          type: ["integer", "null"]
        amountReceived:
          type: string
        currency:
          type: string
        paymentMethod:
          type: string
        receivedDate:
          type: string
        notes:
          type: ["string", "null"]
        tdsApplicable:
          type: boolean
        tdsAmount:
          type: ["string", "null"]
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        dealId:
          type: ["integer", "null"]
        title:
          type: string
        category:
          type: string
        amount:
          type: string
        currency:
          type: string
        expenseDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        recurringFrequency:
          type: ["string", "null"]
        vendorOrPayee:
          type: ["string", "null"]
        dealBrandName:
          type: ["string", "null"]
        dealCampaignName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - category
        - amount
        - currency
        - expenseDate
        - isRecurring
        - createdAt
        - updatedAt
    CreateExpenseBody:
      type: object
      properties:
        dealId:
          type: ["integer", "null"]
        title:
          type: string
        category:
          type: string
        amount:
          type: string
        currency:
          type: string
        expenseDate:
          type: string
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        recurringFrequency:
          type: ["string", "null"]
        vendorOrPayee:
          type: ["string", "null"]
      required:
        - title
        - category
        - amount
        - expenseDate
    UpdateExpenseBody:
      type: object
      properties:
        dealId:
          type: ["integer", "null"]
        title:
          type: string
        category:
          type: string
        amount:
          type: string
        currency:
          type: string
        expenseDate:
          type: string
        notes:
          type: ["string", "null"]
        isRecurring:
          type: boolean
        recurringFrequency:
          type: ["string", "null"]
        vendorOrPayee:
          type: ["string", "null"]
...