# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kitna - Money clarity dashboard for creators and influencer managers
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  userId: number;
  /** @nullable */
  dealId?: number | null;
  title: string;
  category: string;
  amount: string;
  currency: string;
  expenseDate: Date;
  /** @nullable */
  notes?: string | null;
  isRecurring: boolean;
  /** @nullable */
  recurringFrequency?: string | null;
  /** @nullable */
  vendorOrPayee?: string | null;
  /** @nullable */
  dealBrandName?: string | null;
  /** @nullable */
  dealCampaignName?: string | null;
  createdAt: Date;
  updatedAt: Date;
}